moved functions to visualization.utils
